feature: improve stable stringify
This is a performance improvement in many cases. For big / deep
objects it will be a minor performance penalty in the case no
options are used.
If options are used (i.e. replacer or spacer) it will always be
faster.

The implementation is side effect free, so it now works 1-to-1 as
JSON.stringify.